Move locale resolver in enrich bundle
