+++ src/Pim/Bundle/CatalogBundle/Resources/config/builders.yml
@@ -21,7 +21,7 @@ services:
             - '@pim_serializer'
             - '@pim_serializer'
             - '@pim_catalog.builder.product'
-            - '@pim_localization.resolver.locale'
+            - '@pim_enrich.resolver.locale'
             - %pim_catalog.entity.product_template.class%
             - %pim_catalog.entity.product.class%
