+++ src/Pim/Bundle/CommentBundle/Resources/config/normalizers.yml
@@ -6,6 +6,6 @@ services:
         class: %pim_serializer.normalizer.comment.class%
         arguments:
             - '@pim_localization.presenter.datetime'
-            - '@pim_localization.resolver.locale'
+            - '@pim_enrich.resolver.locale'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
