+++ src/Pim/Bundle/EnrichBundle/Resources/config/form_subscribers.yml
@@ -50,7 +50,7 @@ services:
         arguments:
             - '@pim_serializer'
             - '@pim_serializer'
-            - '@pim_localization.resolver.locale'
+            - '@pim_enrich.resolver.locale'
 
     pim_enrich.form.subscriber.add_variant_group_template:
         class: %pim_enrich.form.subscriber.add_variant_group_template.class%
