+++ src/Pim/Bundle/LocalizationBundle/Controller/FormatController.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\LocalizationBundle\Controller;
 
 use Akeneo\Component\Localization\Factory\DateFactory;
 use Akeneo\Component\Localization\Localizer\LocalizerInterface;
-use Pim\Component\Localization\LocaleResolver;
+use Pim\Bundle\EnrichBundle\Resolver\LocaleResolver;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 /**
