+++ src/Pim/Bundle/LocalizationBundle/Resources/config/twig.yml
@@ -12,16 +12,16 @@ services:
     pim_localization.twig.attribute_option_extension:
         class: %pim_localization.twig.attribute_option_extension.class%
         arguments:
-            - '@pim_localization.presenter.registry'
-            - '@pim_localization.resolver.locale'
+            - '@pim_catalog.localization.presenter.registry'
+            - '@pim_enrich.resolver.locale'
         tags:
             - { name: twig.extension }
 
     pim_localization.twig.attribute_extension:
         class: %pim_localization.twig.attribute_extension.class%
         arguments:
-            - '@pim_localization.presenter.date'
-            - '@pim_localization.presenter.datetime'
-            - '@pim_localization.resolver.locale'
+            - '@pim_catalog.localization.presenter.date'
+            - '@pim_catalog.localization.presenter.datetime'
+            - '@pim_enrich.resolver.locale'
         tags:
             - { name: twig.extension }
