+++ src/Pim/Bundle/UIBundle/Resources/config/forms.yml
@@ -29,16 +29,18 @@ services:
     pim_ui.form.type.number:
         class: %pim_ui.form.type.number.class%
         arguments:
-            - '@pim_localization.localizer.number'
-            - '@pim_localization.resolver.locale'
-            - '@pim_localization.validator.constraint.number_format'
+            - '@pim_catalog.localization.localizer.number'
+            - '@pim_enrich.resolver.locale'
+            - '@pim_catalog.localization.validator.constraint.number_format'
+            - '@pim_catalog.localization.factory.number'
         tags:
             - { name: form.type, alias: pim_number }
 
     pim_ui.form.type.date:
         class: %pim_ui.form.type.date.class%
         arguments:
-            - '@pim_localization.resolver.locale'
+            - '@pim_enrich.resolver.locale'
+            - '@pim_catalog.localization.factory.date'
         tags:
             - { name: form.type, alias: pim_date }
